+++ guppy/src/graph/mod.rs
@@ -11,7 +11,6 @@ use petgraph::prelude::*;
 use std::fmt;
 
 mod build;
-mod build_path_out_of_root_helper;
 mod build_targets;
 pub mod cargo;
 mod cycles;
